First frontend layout changes
Created components like the page, the navigation bar, and the 'META' which contains general things like the logo and some css for the whole site. Every page contains the meta and the header so that the whole layout is consistent. The Nav is contained in the header.

home and sell pages changed to reflect these changes as well.